@@ -14,7 +14,11 @@
     "@babel/parser": "^7.14.6",
     "@babel/preset-env": "^7.18.10",
     "@babel/traverse": "^7.14.5",
+    "@types/node": "^20.12.7",
+    "@types/webpack": "^5.28.5",
     "babel-loader": "^8.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.10.0"
@@ -26,7 +30,7 @@
     "query-string": "^8.1.0",
     "react": "^18.2.0",
     "taro-ui": "^3.3.0",
-    "webpack": "^5.84.0",
-    "webpack-cli": "^5.1.1"
+    "webpack": "^5.91.0",
+    "webpack-cli": "^5.1.4"
   }
 }
