@@ -1,6 +1,5 @@
-const { RawSource } = require('webpack').sources
-
-class testPlugin {
+import * as webpack from 'webpack';
+export default class testPlugin {
   // 在插件函数的 prototype 上定义一个 `apply` 方法，以 compiler 为参数。
   apply(compiler) {
     // 指定一个挂载到 webpack 自身的事件钩子。
@@ -10,7 +9,7 @@ class testPlugin {
         // 用 webpack 提供的插件 API 处理构建过程
         compilation.emitAsset(
           'extraFile.json',
-          new RawSource(
+          new webpack.sources.RawSource(
             JSON.stringify({
               x: 1,
               y: 2,
@@ -23,5 +22,3 @@ class testPlugin {
     )
   }
 }
-
-module.exports = testPlugin
